WindowsVersion removed, font changes
WindowsVersion was there mainly to detect XP so that we don't attempt to display certain scripts that were unsupported. That is all removed in 4.1. Also switched the "control font", the one that is used in dialogs, etc., from our bundled fonts to those in the various language packs in Win 10/11.